+++ Runtime/Scripts/Tezos/Wallet/IWalletProvider.cs
@@ -15,7 +15,7 @@ namespace Scripts.Tezos.Wallet
         /// Makes a call to connect with a wallet
         /// <param name="withRedirectToWallet">Should we open wallet app on mobiles after connect?</param>
         /// </summary>
-        void Connect(bool withRedirectToWallet = true);
+        void Connect(WalletProviderType walletProvider, bool withRedirectToWallet = true);
 
         /// <summary>
         /// Unpair with wallet and disconnect