+++ Runtime/Scripts/Tezos/Wallet/WalletProviderType.cs
@@ -1,0 +1,8 @@
+﻿namespace Scripts.Tezos.Wallet
+{
+    public enum WalletProviderType
+    {
+        beacon,
+        kukai
+    }
+}